voice-call(msteams): extract TTS playback into msteams-tts-playback.ts
Move playTts's synthesis + expression/viseme cues + the 20ms frame streamer
(streamPcmFrames, including the closed-socket delivery abort) out of the
provider into a self-contained msteams-tts-playback.ts. The provider's playTts
keeps the state lookup + ttsProvider check and delegates. No behavior change.

Provider 1290 -> 1190 LOC; all playback-only constants/imports moved with it.
